refactor(substrate): split chassis builder into focused modules
